Build a contextmenu in the simplereportpanle to edit the comment of a reportline in a extra dialog.
Now comments are shown by tooltip on mouseover.

added a new methode to reportAgent for updating the comment of 
an existing reportline.

Select reportline lisener works with keyboard shortcuts. 
moved the selecte price lable form the right side of the window to
the top next to the searchbar.
Task #DWOSS-81 - Create a Comment Editor for the ReportLine