Added delet comment into context menu, with a dialog for warning the user that deleting is permanent.
make the deletion of a comment in the edit comment dialog possible.
A small warning get triggert, if the new comment is shorter than 3.
The (OK) of the warning get saved.

the focus of the edit comment dialog get set to the end of the textarea.
Task #DWOSS-111 - ReportLine Kommentar Editor erweitern